Refactor exception mappers into single package (instead of split)
