+++ src/main/java/io/stargate/sgv2/jsonapi/exception/mappers/GenericExceptionMapper.java
@@ -1,8 +1,7 @@
-package io.stargate.sgv2.jsonapi.api.exception;
+package io.stargate.sgv2.jsonapi.exception.mappers;
 
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
 import io.stargate.sgv2.jsonapi.api.model.command.CommandResult;
-import io.stargate.sgv2.jsonapi.exception.mappers.ThrowableCommandResultSupplier;
 import org.jboss.resteasy.reactive.RestResponse;
 import org.jboss.resteasy.reactive.server.ServerExceptionMapper;
